racing_api/requirements.txt
Flask>=3.0.0,<4.0.0
Flask-Cors>=4.0.0,<5.0.0
python-dotenv>=1.0.0,<2.0.0
gunicorn>=22.0.0,<23.0.0
pydantic>=2.6.0,<3.0.0
pytest>=8.0.0,<9.0.0
